Refactor and optimize code structure
