+++ twitter-replica-backend/src/controllers/index.js
@@ -1,0 +1,2 @@
+exports.userController = require('./user.controller');
+exports.postController = require('./post.controller');
