+++ twitter-replica-backend/src/middleware/mediaUpload.js
@@ -1,0 +1,6 @@
+const multer = require('multer');
+const storage = require('../configs/mediaConfig');
+
+const upload = multer({ storage }).single('media');
+
+module.exports = upload;
