+++ twitter-replica-backend/src/routes/api/index.js
@@ -2,6 +2,6 @@ const router = require('express').Router();
 
 // Inject individual routes
 router.use('/', require('./user'));
-router.use('/user/:usernamePrefix/posts', require('./posts'));
+router.use('/user/:usernamePrefix/posts', require('./post'));
 
 module.exports = router;
