--- twitter-replica-backend/src/routes/api/posts.js
@@ -1,127 +1,0 @@
-const asyncHandler = require('express-async-handler');
-const router = require('express').Router({ mergeParams: true });
-const mongoose = require('mongoose');
-const properties = require('../../properties');
-const uploadImage = require('../../configs/mediaConfig');
-const auth = require('../auth');
-
-const Post = mongoose.model('Post');
-
-router.get(
-  '/:postId',
-  asyncHandler(async (req, res) => {
-    try {
-      const result = await Post.findById(req.params.postId);
-      res.json(result).send();
-    } catch (err) {
-      res.status(500).json({ message: 'Could not find post' }).send();
-    }
-  })
-);
-
-router.get(
-  '/',
-  asyncHandler(async (req, res) => {
-    const pageSize = +req.query.pageSize;
-    const currentPage = +req.query.currentPage;
-    let posts;
-    try {
-      if (pageSize && typeof currentPage !== 'undefined') {
-        posts = await Post.find({
-          creatorUsernamePrefix: req.params.usernamePrefix,
-        })
-          .skip(pageSize * currentPage)
-          .limit(pageSize);
-      } else {
-        posts = await Post.find({
-          creatorUsernamePrefix: req.params.usernamePrefix,
-        });
-      }
-      const postCount = await Post.countDocuments({
-        creatorUsernamePrefix: req.params.usernamePrefix,
-      });
-      res.json({ posts, totalPostsCount: postCount }).send();
-    } catch (err) {
-      res.status(500).json({ message: 'Could not fetch posts' }).send();
-    }
-  })
-);
-
-router.post(
-  '/',
-  auth,
-  uploadImage,
-  asyncHandler(async (req, res) => {
-    const url = `${req.protocol}://${req.get('host')}`;
-    const post = new Post({
-      content: req.body.content,
-      mediaPath: url + properties.mediaPath + req.file.filename,
-      creatorId: req.userData.id,
-      creatorUsernamePrefix: req.userData.usernamePrefix,
-    });
-    try {
-      const ret = await post.save();
-      res.json(ret).send();
-    } catch (err) {
-      res.status(500).json({ message: 'Could not add post' }).send();
-    }
-  })
-);
-
-router.put(
-  '/:postId',
-  auth,
-  uploadImage,
-  asyncHandler(async (req, res) => {
-    let currentMedia = req.body.mediaPath;
-
-    if (req.file) {
-      const url = `${req.protocol}://${req.get('host')}`;
-      currentMedia = url + properties.mediaPath + req.file.filename;
-    }
-
-    const post = new Post({
-      _id: req.body.id,
-      content: req.body.content,
-      mediaPath: currentMedia,
-      creatorId: req.userData.id,
-      creatorUsernamePrefix: req.userData.usernamePrefix,
-    });
-
-    try {
-      const result = await Post.updateOne(
-        { _id: req.params.postId, creatorId: req.userData.id },
-        post
-      );
-      if (result.nModified <= 0) {
-        res.status(401).json({ message: 'Not authorized' }).send();
-      }
-      res.json(post).send();
-    } catch (err) {
-      res.status(500).json({ message: 'Could not update post' }).send();
-    }
-  })
-);
-
-router.delete(
-  '/:postId',
-  auth,
-  asyncHandler(async (req, res) => {
-    try {
-      const result = await Post.deleteOne({
-        _id: req.params.postId,
-        creatorId: req.userData.id,
-      });
-      if (result.n <= 0) {
-        res.status(401).json({ message: 'Not authorized' }).send();
-      }
-
-      const totalPosts = await Post.estimatedDocumentCount();
-      res.json({ totalPosts }).send();
-    } catch (err) {
-      res.status(401).json({ message: 'Could not remove post' }).send();
-    }
-  })
-);
-
-module.exports = router;
