+++ twitter-replica-ui/libs/twitter-auth/src/lib/twitter-auth.module.ts
@@ -2,11 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { ReactiveFormsModule } from '@angular/forms';
 import { RouterModule, Route } from '@angular/router';
-import { MatCardModule } from '@angular/material/card';
-import { MatFormFieldModule } from '@angular/material/form-field';
-import { MatInputModule } from '@angular/material/input';
-import { MatButtonModule } from '@angular/material/button';
-
+import { MaterialModule } from '../../../material.module';
 import { TwitterLoginFormComponent } from './login/login.component';
 import { TwitterSingupFormComponent } from './signup/signup.component';
 
@@ -26,15 +22,7 @@ export const TwitterSignupRoutes: Route[] = [
 
 @NgModule({
   declarations: [TwitterLoginFormComponent, TwitterSingupFormComponent],
-  imports: [
-    CommonModule,
-    ReactiveFormsModule,
-    RouterModule,
-    MatCardModule,
-    MatFormFieldModule,
-    MatInputModule,
-    MatButtonModule,
-  ],
+  imports: [CommonModule, MaterialModule, ReactiveFormsModule, RouterModule],
   exports: [],
 })
 export class TwitterAuthModule {}
