+++ twitter-replica-ui/libs/twitter-user/src/lib/twitter-user.module.ts
@@ -1,12 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterModule, Route } from '@angular/router';
-import { MatInputModule } from '@angular/material/input';
-import { MatCardModule } from '@angular/material/card';
-import { MatButtonModule } from '@angular/material/button';
-import { MatExpansionModule } from '@angular/material/expansion';
-import { MatPaginatorModule } from '@angular/material/paginator';
-import { MatIconModule } from '@angular/material/icon';
+import { MaterialModule } from '../../../material.module';
 
 import { TwitterUserHomepageComponent } from './user-homepage/user-homepage.component';
 import { TwitterPostModule } from '@twitter-replica/post';
@@ -19,17 +14,7 @@ export const TwitterUserRoutes: Route[] = [
 ];
 
 @NgModule({
-  imports: [
-    CommonModule,
-    RouterModule,
-    TwitterPostModule,
-    MatInputModule,
-    MatCardModule,
-    MatButtonModule,
-    MatIconModule,
-    MatExpansionModule,
-    MatPaginatorModule,
-  ],
+  imports: [CommonModule, RouterModule, MaterialModule, TwitterPostModule],
   declarations: [TwitterUserHomepageComponent],
   exports: [TwitterUserHomepageComponent],
 })
